refactor: builder -> scaneer
