+++ pkg/cmd/scanner/scheme_test.go
@@ -1,4 +1,4 @@
-package resource
+package scanner
 
 import (
 	"testing"
