+++ pkg/cmd/scanner/yamljson.go
@@ -1,4 +1,4 @@
-package resource
+package scanner
 
 import (
 	"encoding/json"
